server/py/requirements.txt
mss
Pillow
pynput
screeninfo
clipboard
py2exe